requirements.txt
beautifulsoup4==4.12.2
httpx==0.24.0
Pillow==9.5.0
URLObject==2.4.3
